Moved filtering actions to a different file
